backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["halilagin <halil.agin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.115.11"
uvicorn = "0.24.0"
sqlalchemy = "2.0.23"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-jose = {version = "3.3.0", extras = ["cryptography"]}
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
python-multipart = "0.0.6"
python-dotenv = "1.0.0"
pydantic-settings = "2.1.0"
bcrypt = ">=4.0.0"
pytest = "^8.3.5"
httpx = "^0.28.1"
starlette = "^0.46.0"
pytest-asyncio = "^0.25.3"
reportlab = "^4.3.1"
pypdf2 = "^3.0.1"
pypdf = "^5.3.1"
endesive = "^2.18.3"
pyjwt = "^2.10.1"
stripe = "^11.6.0"
psycopg2-binary = "^2.9.10"
pycryptodome = "^3.22.0"
jinja2 = "^3.1.4"
pillow = "^11.2.1"
aiosmtplib = "^3.0.1"
apscheduler = "^3.10.4"
pytz = "^2025.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
